% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kalmanFilter}
\alias{kalmanFilter}
\title{Function For Predicting States using Kalman Filter}
\usage{
kalmanFilter(z, x, f, p, h, q, r)
}
\arguments{
\item{z}{Best estimate of measurements}

\item{x}{Initial state}

\item{f}{Prediction matrix used to obtain next state}

\item{p}{Error covariance matrix of the state estimate}

\item{h}{Matrix to model the signal}

\item{q}{Covariance matrix to model uncertainity associated with untracked influences}

\item{r}{Covariance matrix to model uncertainity eg. sensor noise}
}
\value{
Updated states/measurements
}
\description{
The function allows you input the parameter matrices of
the Kalman Filter and returns a matrix containing
predicted states/mesurements using given parameters.
}
\keyword{KalmanFilter}
