\name{KFilter-package}
\alias{KFilter-package}
\alias{KFilter}
\docType{package}
\title{
\packageTitle{KFilter}
}
\description{
The package contains a sole functoin kalmanFilter which takes 
parameters of a Kalman Filter as input and outputs best predicted 
states.
}
\author{
Hrushikesh Turkhade, 
Maintainer: Hrushikesh Turkahde<hrushi94205@gmail.com>
}

